# PERF/REF: groupby sample