… asv_bench/benchmarks/groupby.py
… doc/source/whatsnew/v1.4.0.rst
… pandas/core/sample.py
… pandas/tests/frame/methods/test_sample.py
… pandas/tests/groupby/test_sample.py